requirements.txt
fastapi
uvicorn
tensorflow
python-multipart
opencv-python
websockets
pillow
requests